chore: refactor components imports to default imports
