+++ src/components/Hero/index.js
