+++ src/pages/poland-form/index.js
@@ -1,5 +1,5 @@
-import Layout from "../../components/Layout/Layout";
-import Hero from "../../components/Hero/Hero";
+import Layout from "../../components/Layout";
+import Hero from "../../components/Hero";
 
 import { useTranslation } from "react-i18next";
